+++ CHANGELOG.md
@@ -10,6 +10,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -
 
+## Release - v1.15.2 - 26-05-2026
+
+- Rolledback remove MP optional payer email [PR#192](https://github.com/silvioubaldino/personal-finance/pull/192)
+- Allowed retry coupon checkout when coupon is pending [PR#191](https://github.com/silvioubaldino/personal-finance/pull/191)
+
+## Release - v1.15.1 - 26-05-2026
+
+- Removed MP optional payer email [PR#190](https://github.com/silvioubaldino/personal-finance/pull/190)
+
 ## Release - v1.15.0 - 26-05-2026
 - Fixed 500 on credit card movement when default wallet is nil [PR#188](https://github.com/silvioubaldino/personal-finance/pull/188)
 - Created coupon entity [PR#187](https://github.com/silvioubaldino/personal-finance/pull/187)